Speed and stability improvements
Added protection to streams to gracefully handle concurrent reads.

When mounting as files, a short performance test is run to see how long it takes to extract a small file. If fast, then 7z.exe is used, if slow then 7zFM.exe is used.

WaitForDataStream allows the data coming back from 7z.exe to be immediately provided to the reader, instead of needing to be fully extracted first.

Fixed bug in GZipStreamSeekable's read recommender, which was causing an endless loop in CachedStream.